DatLA add new source code for function 5.4.1 show list of notification and delete notification. 9h47 03/02

+++ C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/dto/notification/NotificationDeleteDto.java
@@ -6,6 +6,6 @@ package com.c0722g1repobe.dto.notification;
  * For: to show list notification to delete
  */
 public interface NotificationDeleteDto {
-    Long getIdNotification();
+    Integer getIdNotification();
     String getTitle();
 }
